[ 🔧 | #52 | refactor ] MainPage 리팩토링 완료
