+++ src/pages/MainPage/data/mock.js
